pyhub/mydjango/requirements.txt
django~=5.2.0
django-environ
pyhub-llm[all]

django-bootstrap5

django-debug-toolbar
django-extensions
